+++ lib/roby/actions.rb
@@ -7,6 +7,7 @@ end
 require 'roby/actions/models/action'
 require 'roby/actions/models/method_action'
 require 'roby/actions/models/coordination_action'
+require 'roby/actions/models/interface_base'
 require 'roby/actions/models/interface'
 require 'roby/actions/models/library'
