+++ lib/roby/actions/models/library.rb
@@ -3,7 +3,7 @@ module Roby
         module Models
             # Metamodel for action libraries
             class Library < Module
-                include Models::Interface
+                include Models::InterfaceBase
                 include MetaRuby::ModelAsModule
                 include MetaRuby::Registration
             end
